+++ packages/address/src/address.ts
@@ -64,6 +64,22 @@ export default class Address extends AbstractAddress {
   }
 
   /**
+   * Parses this Address value
+   * @returns a string address in Bech32m Format
+   */
+  toJSON(): string {
+    return this.toString();
+  }
+
+  /**
+   * Returns the value of this Address value
+   * @returns a string address in Bech32m Format
+   */
+  valueOf(): string {
+    return this.toString();
+  }
+
+  /**
    * Compare this Address value to another for direct equality
    * @param other - the other address to compare against
    * @returns true if addresses are equal