+++ packages/contract/scripts/build-multicall.sh
@@ -6,7 +6,7 @@ FUELS_DEBUG=${FUELS_DEBUG:-}
 SCRIPT_DIR="src/contracts/multicall"
 BIN_DIR="$SCRIPT_DIR/out/debug"
 
-pnpm forc build -p $SCRIPT_DIR --print-finalized-asm
+pnpm forc build -p $SCRIPT_DIR --print-finalized-asm --generate-logged-types
 pnpm exec ts-node scripts/process-multicall.ts
 if [[ -n "$FUELS_DEBUG" ]]; then
   pnpm forc parse-bytecode "$BIN_DIR/multicall.bin" > "$BIN_DIR/multicall.txt"