+++ packages/contract/src/index.ts
@@ -3,4 +3,8 @@ export { default as ContractFactory } from './contracts/contract-factory';
 export { default as Contract } from './contracts/contract';
 export { FunctionInvocationScope } from './contracts/functions/invocation-scope';
 export { MultiCallInvocationScope } from './contracts/functions/multicall-scope';
+export {
+  InvocationResult,
+  FunctionInvocationResult,
+} from './contracts/functions/invocation-results';
 export * as ContractUtils from './util';