+++ packages/fuel-gauge/src/utils.ts
@@ -1,0 +1,55 @@
+import { readFileSync } from 'fs';
+import type { Interface, JsonAbi, Wallet, Contract, BytesLike } from 'fuels';
+import { NativeAssetId, Provider, TestUtils, ContractFactory } from 'fuels';
+import { join } from 'path';
+
+let contractInstance: Contract;
+const deployContract = async (factory: ContractFactory, useCache: boolean = true) => {
+  if (contractInstance && useCache) return contractInstance;
+  contractInstance = await factory.deployContract();
+  return contractInstance;
+};
+
+let walletInstance: Wallet;
+const createWallet = async () => {
+  if (walletInstance) return walletInstance;
+  const provider = new Provider('http://127.0.0.1:4000/graphql');
+  walletInstance = await TestUtils.generateTestWallet(provider, [
+    [5_000_000, NativeAssetId],
+    [5_000_000, '0x0101010101010101010101010101010101010101010101010101010101010101'],
+  ]);
+  return walletInstance;
+};
+
+export type SetupConfig = {
+  contractBytecode: BytesLike;
+  abi: JsonAbi | Interface;
+  cache?: boolean;
+};
+
+export const setup = async ({ contractBytecode, abi, cache }: SetupConfig) => {
+  // Create wallet
+  const wallet = await createWallet();
+  const factory = new ContractFactory(contractBytecode, abi, wallet);
+  const contract = await deployContract(factory, cache);
+  return contract;
+};
+
+export const createSetupConfig =
+  (defaultConfig: SetupConfig) => async (config?: Partial<SetupConfig>) =>
+    setup({
+      contractBytecode: defaultConfig.contractBytecode,
+      abi: defaultConfig.abi,
+      ...config,
+    });
+
+const getFullPath = (contractName: string, next: (fullPath: string) => () => Promise<Contract>) =>
+  next(join(__dirname, `../test-projects/${contractName}/out/debug/${contractName}`));
+
+export const getSetupContract = (contractName: string) =>
+  getFullPath(contractName, (fullPath: string) =>
+    createSetupConfig({
+      contractBytecode: readFileSync(`${fullPath}.bin`),
+      abi: JSON.parse(readFileSync(`${fullPath}-abi.json`, 'utf8')),
+    })
+  );