+++ packages/script/scripts/build-call-test-script.sh
@@ -6,7 +6,7 @@ FUELS_DEBUG=${FUELS_DEBUG:-}
 SCRIPT_DIR="src/call-test-script"
 BIN_DIR="$SCRIPT_DIR/out/debug"
 
-pnpm forc build -p $SCRIPT_DIR --print-finalized-asm
+pnpm forc build -p $SCRIPT_DIR --print-finalized-asm --generate-logged-types
 if [[ -n "$FUELS_DEBUG" ]]; then
   pnpm forc parse-bytecode "$BIN_DIR/call-test-script.bin" > "$BIN_DIR/call-test-script.txt"
 fi