+++ packages/transactions/CHANGELOG.md
@@ -1,5 +1,50 @@
 # Change Log
 
+## 0.17.0
+
+### Minor Changes
+
+- [#517](https://github.com/FuelLabs/fuels-ts/pull/517) [`6403076`](https://github.com/FuelLabs/fuels-ts/commit/6403076bb9fce9055b436596e23713b0e7909d87) Thanks [@QuinnLee](https://github.com/QuinnLee)! - Parse Logs and Log Data
+
+### Patch Changes
+
+- [#515](https://github.com/FuelLabs/fuels-ts/pull/515) [`fa83fcd`](https://github.com/FuelLabs/fuels-ts/commit/fa83fcd0c90ddb95bc397ab2675a5ad759b94f82) Thanks [@camsjams](https://github.com/camsjams)! - Added message support
+
+- Updated dependencies [[`f106a78`](https://github.com/FuelLabs/fuels-ts/commit/f106a78e816045e3bdb6bff0b9bceec871009091), [`658b065`](https://github.com/FuelLabs/fuels-ts/commit/658b06538389a6ad3310a739a1bf60311c1e3343), [`6403076`](https://github.com/FuelLabs/fuels-ts/commit/6403076bb9fce9055b436596e23713b0e7909d87)]:
+  - @fuel-ts/math@0.17.0
+  - @fuel-ts/abi-coder@0.17.0
+  - @fuel-ts/constants@0.17.0
+
+## 0.16.0
+
+### Minor Changes
+
+- [#508](https://github.com/FuelLabs/fuels-ts/pull/508) [`27224b9`](https://github.com/FuelLabs/fuels-ts/commit/27224b997a4ec86473fc19868550c788638fa2ce) Thanks [@luizstacio](https://github.com/luizstacio)! - Add array-coder on exports
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fuel-ts/abi-coder@0.16.0
+  - @fuel-ts/constants@0.16.0
+  - @fuel-ts/math@0.16.0
+
+## 0.15.0
+
+### Minor Changes
+
+- [#495](https://github.com/FuelLabs/fuels-ts/pull/495) [`f3c7273`](https://github.com/FuelLabs/fuels-ts/commit/f3c7273d946979e628b178ba808b8fc1598105bb) Thanks [@luizstacio](https://github.com/luizstacio)! - Fix exports and imports
+
+- [#494](https://github.com/FuelLabs/fuels-ts/pull/494) [`9d0ad53`](https://github.com/FuelLabs/fuels-ts/commit/9d0ad5392b2dae83b13041999435c08e07e935a3) Thanks [@QuinnLee](https://github.com/QuinnLee)! - fix transaction decode offset
+
+### Patch Changes
+
+- [#468](https://github.com/FuelLabs/fuels-ts/pull/468) [`7ad3d79`](https://github.com/FuelLabs/fuels-ts/commit/7ad3d79bf7a1db766912a7b3d52e4fa2e550af56) Thanks [@LuizAsFight](https://github.com/LuizAsFight)! - Refactor to use bn.js instead of bigint.
+
+- Updated dependencies [[`63583aa`](https://github.com/FuelLabs/fuels-ts/commit/63583aa6e8b5b5417bdc0c0ae3bc15eec7735e43), [`7ad3d79`](https://github.com/FuelLabs/fuels-ts/commit/7ad3d79bf7a1db766912a7b3d52e4fa2e550af56), [`f3c7273`](https://github.com/FuelLabs/fuels-ts/commit/f3c7273d946979e628b178ba808b8fc1598105bb)]:
+  - @fuel-ts/abi-coder@0.15.0
+  - @fuel-ts/constants@0.15.0
+  - @fuel-ts/math@0.15.0
+
 ## 0.14.0
 
 ### Patch Changes